requirements.txt
Django==1.9
djangorestframework==3.3.2
Pygments==2.0.2
wheel==0.24.0
